Refactor file names (#12)
* refactor file names

* reverting cmake unnessecary changes.
