+++ demos/TicTacToe/Classes/tic_tac_toe_layer.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "TicTacToeLayer.h"
+#include "tic_tac_toe_layer.h"
 
 USING_NS_CC;
