+++ demos/TicTacToe/Classes/tic_tac_toe_layer.h
@@ -27,8 +27,8 @@
 #include <string>
 #include <unordered_set>
 
-#include "TicTacToeScene.h"
 #include "cocos2d.h"
+#include "tic_tac_toe_scene.h"
 #include "util.h"
 
 using cocos2d::Director;
