+++ demos/TicTacToe/Classes/tic_tac_toe_scene.h
@@ -15,9 +15,9 @@
 #ifndef TICTACTOE_DEMO_CLASSES_TICTACTOE_SCENE_H_
 #define TICTACTOE_DEMO_CLASSES_TICTACTOE_SCENE_H_
 
-#include "MainMenuScene.h"
-#include "TicTacToeLayer.h"
 #include "cocos2d.h"
+#include "main_menu_scene.h"
+#include "tic_tac_toe_layer.h"
 #include "util.h"
 
 class TicTacToe : public cocos2d::Layer {
